Ensure that way too large custom cursors can't be created

+++ examples/custom_cursors.rs
@@ -13,7 +13,7 @@ fn decode_cursor(bytes: &[u8]) -> CustomCursor {
     let img = image::load_from_memory(bytes).unwrap().to_rgba8();
     let samples = img.into_flat_samples();
     let (_, w, h) = samples.extents();
-    let (w, h) = (w as u32, h as u32);
+    let (w, h) = (w as u16, h as u16);
     CustomCursor::from_rgba(samples.samples, w, h, w / 2, h / 2).unwrap()
 }
